pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cos-python-sdk-v5>=1.9.38",
    "fastapi>=0.119.0",
    "pydantic[email]>=2.12.2",
    "pydantic-settings>=2.11.0",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.44",
    "tencentcloud-sdk-python-sts>=3.0.1459",
    "uvicorn>=0.37.0",
    "asyncmy>=0.2.10",
    "funboost>=50.4",
    "ulid-py>=1.1.0",
    "alembic>=1.17.0",
    "httpx>=0.28.1",
    "python-jose[cryptography]>=3.5.0",
    "hypercorn>=0.17.3",
    "asyncio>=4.0.0",
    "psutil>=7.1.1",
    "passlib[bcrypt]>=1.7.4",
    "argon2-cffi>=25.1.0",
]
